requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
# mysqlclient==2.2.0
python-decouple==3.8
django-filter==23.3
Pillow==10.1.0
python-dateutil==2.8.2
google-auth-oauthlib==1.2.2
google-auth==2.40.3
requests==2.32.4
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
daphne==4.0.0
django-storages==1.14.2
boto3==1.35.71